# docs/swagger.yaml
basePath: /
definitions:
...
  events.ActorType:
    enum:
    - OWNER_USER
    - DELEGATE_USER
    - EXTERNAL_SYSTEM
    type: string
    x-enum-varnames:
    - ActorTypeOwnerUser
    - ActorTypeDelegateUser
    - ActorTypeExternalSystem
  events.EventStatus:
    enum:
    - active
    - voided
    type: string
    x-enum-varnames:
    - EventStatusActive
    - EventStatusVoided
  events.EventType:
    enum:
    - PROFILE_UPDATED
    - WEIGHT_RECORDED
    - NOTE
    - MEDICAL_VISIT
    - MEDICATION_PRESCRIBED
    - VACCINE
    - DEWORMING
    - FLEA_TREATMENT
    - BATH
    - ATTACHMENT_ADDED
    type: string
    x-enum-varnames:
    - EventTypeProfileUpdated
    - EventTypeWeightRecorded
    - EventTypeNote
    - EventTypeMedicalVisit
    - EventTypeMedicationPresc
    - EventTypeVaccine
    - EventTypeDeworming
    - EventTypeFleaTreatment
    - EventTypeBath
    - EventTypeAttachmentAdded
  events.Source:
    enum:
    - manual
    - smartpet
    - integration
    type: string
    x-enum-varnames:
    - SourceManual
    - SourceSmartPet
    - SourceIntegration
  events.Visibility:
    enum:
    - private
    - shared_with_delegates
    type: string
    x-enum-varnames:
    - VisibilityPrivate
    - VisibilityShared
  events.createEventRequest:
    properties:
      notes:
        type: string
      occurred_at:
        description: RFC3339
        type: string
      source:
        allOf:
        - $ref: '#/definitions/events.Source'
        description: opcional
      title:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/events.EventType'
        enum:
        - NOTE
        - MEDICAL_VISIT
        - VACCINE
        - DEWORMING
        - BATH
        - PROFILE_UPDATED
        - WEIGHT_RECORDED
        - MEDICATION_PRESCRIBED
        - FLEA_TREATMENT
        - ATTACHMENT_ADDED
      visibility:
        allOf:
        - $ref: '#/definitions/events.Visibility'
        description: opcional
    type: object
  events.eventResponse:
    properties:
      actor_id:
        type: string
      actor_type:
        $ref: '#/definitions/events.ActorType'
      id:
        type: string
      notes:
        type: string
      occurred_at:
        type: string
      pet_id:
        type: string
      recorded_at:
        type: string
      source:
        $ref: '#/definitions/events.Source'
      status:
        $ref: '#/definitions/events.EventStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/events.EventType'
      visibility:
        $ref: '#/definitions/events.Visibility'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@odyssoft.com
    name: API Support
    url: https://odyssoft.com/support
  description: API REST para gestión de historiales clínicos de mascotas con sistema
    de delegación de acceso.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pet Clinical History API
  version: "1.0"
paths:
...
  /pets/{petID}/events:
    get:
      consumes:
      - application/json
      description: 'Lista los eventos clínicos de una mascota. El dueño siempre puede
        verlos. Un delegado necesita un grant activo con scope `events:read`. Autenticación:
        `X-Debug-User-ID` (dev) o `Authorization: Bearer <token>` (prod). Permite
        filtrar por tipos, rango de fechas y texto.'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      - description: ID de la mascota
        in: path
        name: petID
        required: true
        type: string
      - description: Máximo de eventos a devolver (1-200). Por defecto 50
        in: query
        name: limit
        type: integer
      - description: 'Lista CSV de tipos de evento a incluir (ej: MEDICAL_VISIT,BATH)'
        in: query
        name: types
        type: string
      - description: Fecha/hora mínima occurred_at (RFC3339)
        in: query
        name: from
        type: string
      - description: Fecha/hora máxima occurred_at (RFC3339)
        in: query
        name: to
        type: string
      - description: Texto de búsqueda libre en título/notas
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/events.eventResponse'
            type: array
        "400":
          description: Parámetros de filtro inválidos
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
        "404":
          description: pet not found
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Listar eventos de una mascota
      tags:
      - events
    post:
      consumes:
      - application/json
      description: 'Crea un nuevo evento clínico para la mascota indicada. El dueño
        siempre puede crear eventos. Un delegado necesita un grant activo con scope
        `events:create`. Autenticación: `X-Debug-User-ID` (dev) o `Authorization:
        Bearer <token>` (prod).'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      - description: ID de la mascota
        in: path
        name: petID
        required: true
        type: string
      - description: Datos del evento; occurred_at en formato RFC3339
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/events.createEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/events.eventResponse'
        "400":
          description: invalid json / occurred_at inválido / reglas de negocio
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
        "404":
          description: pet not found
          schema:
            type: string
      summary: Crear evento de mascota
      tags:
      - events
  /pets/{petID}/events/{eventID}/void:
    post:
      consumes:
      - application/json
      description: 'Anula un evento existente de la mascota. El dueño siempre puede
        anular. Un delegado necesita un grant activo con scope `events:void`. Autenticación:
        `X-Debug-User-ID` (dev) o `Authorization: Bearer <token>` (prod).'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      - description: ID de la mascota
        in: path
        name: petID
        required: true
        type: string
      - description: ID del evento
        in: path
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventResponse'
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
        "404":
          description: event not found
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Anular (void) un evento
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT obtenido de Odin-IAM. Formato: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
  DebugUserID:
    description: Solo en modo dev. Permite simular autenticación sin token real.
    in: header
    name: X-Debug-User-ID
    type: apiKey
swagger: "2.0"